also move cache_streaming into cache_streaming module
